Refactor for better ES6 named exports support
