+++ lib/Server.js
@@ -1,0 +1,55 @@
+'use strict'
+
+var crypto = require('crypto')
+var utils = require('./utils')
+
+var Server = function (options) {
+  this.host = options.host || 'https://sessionserver.mojang.com'
+}
+
+/**
+ * Client's Mojang handshake call
+ * See http://wiki.vg/Protocol_Encryption#Client
+ * @param  {String}   token        Client's accessToken
+ * @param  {String}   profile      Client's selectedProfile
+ * @param  {String}   serverid     ASCII encoding of the server ID
+ * @param  {String}   sharedsecret Server's secret string
+ * @param  {String}   serverkey    Server's encoded public key
+ * @param  {Function} cb           (is okay, data returned by server)
+ * @async
+ */
+Server.prototype.join = function (token, profile, serverid, sharedsecret, serverkey, cb) {
+  utils.call(this.host, 'session/minecraft/join', {
+    'accessToken': token,
+    'selectedProfile': profile,
+    'serverId': utils.mcHexDigest(crypto.createHash('sha1').update(serverid).update(sharedsecret).update(serverkey).digest())
+  }, function (err, data) {
+    cb(err, data)
+  })
+}
+
+/**
+ * Server's Mojang handshake call
+ * @param  {String}   username     Client's username, case-sensitive
+ * @param  {String}   serverid     ASCII encoding of the server ID
+ * @param  {String}   sharedsecret Server's secret string
+ * @param  {String}   serverkey    Server's encoded public key
+ * @param  {Function} cb           (is okay, client info)
+ * @async
+ */
+Server.prototype.hasJoined = function (username, serverid, sharedsecret, serverkey, cb) {
+  var hash = utils.mcHexDigest(crypto.createHash('sha1').update(serverid).update(sharedsecret).update(serverkey).digest())
+  utils.request({
+    method: 'GET',
+    uri: this.host + '/session/minecraft/hasJoined?username=' + username + '&serverId=' + hash,
+    json: true
+  }, function (err, resp, body) {
+    if (err || body && body.hasOwnProperty('id')) {
+      cb(err, body)
+    } else {
+      cb(new Error('Failed to verify username!'))
+    }
+  })
+}
+
+module.exports = Server
